Changing ATTR_PG from 32 to 16 bit
Original-Change-Id: I346f10136e9621ea06e381cfad2a2b903cb2bd64
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/29834
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Dev-Ready: Joseph J. McGill <jmcgill@us.ibm.com>

+++ src/import/chips/p9/procedures/hwp/perv/p9_sbe_chiplet_reset.C
@@ -961,7 +961,7 @@ fapi_try_exit:
 static fapi2::ReturnCode p9_sbe_chiplet_reset_mc_setup_cache(
     const fapi2::Target<fapi2::TARGET_TYPE_PERV>& i_target_chiplet)
 {
-    uint32_t l_attr_pg = 0;
+    uint16_t l_attr_pg = 0;
     FAPI_INF("p9_sbe_chiplet_reset_mc_setup_cache: Entering ...");
 
     FAPI_DBG("Reading ATTR_PG");
